Prepare for first launch

+++ src/Services/FileHostingGateway/FileHostingGateway.Application/Commands/UploadVideo/UploadVideoCommand.cs
@@ -4,6 +4,7 @@ using System.Threading;
 using System.Threading.Tasks;
 
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Hosting;
 
 using MediatR;
 
@@ -11,26 +12,27 @@ using FileHostingGateway.Application.Common.Interfaces;
 using FileHostingGateway.Application.Common.Results;
 
 namespace FileHostingGateway.Application.Commands.UploadVideo {
-    public class UploadVideoCommand : IRequest<HandleResult<VideoStreamingInfoDto>> {
+    public class UploadVideoCommand : IRequest<HandleResult<string>> {
         public string FilePath { get; init; }
         public string VimeoProjectId { get; init; }
     }
 
-    public class UploadVideoCommandHandler : IRequestHandler<
-        UploadVideoCommand, HandleResult<VideoStreamingInfoDto>
-    > {
+    public class UploadVideoCommandHandler : IRequestHandler<UploadVideoCommand, HandleResult<string>> {
         private readonly ILogger<UploadVideoCommandHandler> _logger;
+        private readonly IHostEnvironment _hostEnvironment;
         private readonly IVimeoGateway _vimeoGateway;
 
         public UploadVideoCommandHandler(
             ILogger<UploadVideoCommandHandler> logger,
+            IHostEnvironment hostEnvironment,
             IVimeoGateway vimeoGateway
         ) {
             _logger = logger;
+            _hostEnvironment = hostEnvironment;
             _vimeoGateway = vimeoGateway;
         }
 
-        public async Task<HandleResult<VideoStreamingInfoDto>> Handle(
+        public async Task<HandleResult<string>> Handle(
             UploadVideoCommand command, CancellationToken cancellationToken
         ) {
             try {
@@ -38,16 +40,18 @@ namespace FileHostingGateway.Application.Commands.UploadVideo {
                 if (outcome.IsError) {
                     _logger.LogError(outcome.Error.Errors.Values.First().First());
 
-                    return new HandleResult<VideoStreamingInfoDto> {
+                    return new HandleResult<string> {
                         Error = outcome.Error
                     };
                 }
 
-                return new HandleResult<VideoStreamingInfoDto> {
+                return new HandleResult<string> {
                     Data = outcome.Data
                 };
             } finally {
-                File.Delete(command.FilePath);
+                if (!_hostEnvironment.IsDevelopment()) {
+                    File.Delete(command.FilePath);
+                }
             }
         }
     }
